update prediction screen, make top bar transparent, remove all tflite functional

+++ app/src/main/java/com/refanzzzz/palmoildetection/data/response/PredictResponse.kt
@@ -11,7 +11,7 @@ data class PredictItem(
     @SerializedName("label")
     val prediction: String,
     @SerializedName("confidence")
-    val confidence: Double
+    val confidence: Float
 )
